Move EXPECTED_GRADES examples to arithmetic

+++ examples/arithmetic/submissions/partially_accepted/arithmetic_dir/main.cpp
@@ -1,9 +1,10 @@
 #include <bits/stdc++.h>
+// A submission in a directory
 // @EXPECTED_GRADES@ AC AC WA WA
 using namespace std;
 
 int main() {
-	long double a, b, c;
+	double a, b, c;
 	cin >> a >> b >> c;
 	cout << setprecision(18) << fixed << a * b / c << endl;
 }
